Added a color selector
Used the built in html color selector then saved the value ina  variable. That was then used to change the color of the cells. I also moved some things into functions so that the grid will load in when the page loads and so that the slider will have its own event listener.