Finished rainbow mode
I debugged the rainbow mode section on the etch a sketch, I had an if statement to test for if rainbow mode was activated, but forgot to return the function if it was, resulting in it being overridden. Solved that, then removed a . from in front of a class to have a visual toggle.